refactor(PlacementConstraints): move ConstraintsField to component
Relates to DCOS-21167
